requirements.txt
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
requests==2.31.0
Pillow==10.4.0
aiohttp==3.9.3
python-dotenv==1.0.0 